Create v1 version of GC logic
